refactor to have BaseFeature class
